Finish Lab 1
Actually, the idea about the Lab 1 is the same as the Lab 0. Actually
for efficiency, I use tow `vector` to finish this task. The only memory
allocation is happened at the constructor. I don't use the
`unordered_set` to hold the state. Instead, I use `_dirty` vector to map
the state.